ManualControl: fix RC override
